RandomString: explicit float cast keeps int*float as fmul+fimul order
Pre-casting the masked rand() to float forces MSVC6 to multiply the
float constant before the list->numstrings fimul, matching ref's
'fmul DWORD PTR ds:IMM; fimul DWORD PTR [edi+0x4]' instruction order.

Source-commit: 37c7ebdce6b0b8d8c4d08b919128208cad45fea5

+++ botlib/botlib.c
@@ -26226,7 +26226,7 @@ char *__cdecl RandomString(const char *name)
   {
     if ( !strcmp(list->string, name) )
     {
-      n = (int)((rand() & 0x7FFF) * 0.000030518509f * list->numstrings);
+      n = (int)((float)(rand() & 0x7FFF) * 0.000030518509f * list->numstrings);
       rs = list->firstrandomstring;
       while ( rs )
       {
